Shadcn & universal components (#159)
* Add Shadcn

* Add ShadCN button and input

* Create shadcn.md

* Shadcn badge and textarea

* Run prettier

* Refactor inputs

* Move input component

* Refactor button sizes

* wip

* wip

* Move Shadcn components

* Rework colors

* Update globals.scss

* Setup tooltip

* Refactor buttons

* Update tailwind.config.ts

* wip

* Fixing colors

* Update globals.scss

* wip

* Update globals.scss

* wip

* wip

* Add dropdown menu

* Rework theme switch

* Update ThemeSwitch.tsx

* wip

* components/common is now components/ui

* wip

* Rework status labels

* badges

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* Update index.tsx

* wip

* wip

* Update LinkCard.tsx

* wip

* Checkbox and label

* wip

* Implement form labels

* wip

* Update SideModal.tsx

* wip

* wip

* Update EditorLayoutWrapper.tsx

* wip

* Update index.tsx

* wip

* wip

* Update CollectionSchema.tsx

* Update index.tsx

* wip

* wip

* wip

* Update CollectionReferenceList.tsx

* wip

* wip

* wip

* fix infinite rerender in the editor

* fix build

+++ irmin-console/src/components/repository/RepositorySettingsSection.tsx
@@ -2,14 +2,14 @@
 
 import React, { useCallback } from 'react';
 
-import SettingsForm, {
-  FieldConfig,
-} from '@/components/common/form/SettingsForm';
+import SettingsForm, { FieldConfig } from '@/components/ui/form/SettingsForm';
 
 import { useLocale } from '@/context/LocaleContext';
 import { usePopup } from '@/context/PopupContext';
 import { useRepository } from '@/context/RepositoryContext';
 import { useWorkspace } from '@/context/workspace';
+
+import ImmutableWarning from './ImmutableWarning';
 
 /**
  * Repository Settings section component
@@ -129,31 +129,29 @@ const RepositorySettingsSection = () => {
     },
   ];
 
+  if (currentRepository?.is_immutable) {
+    return <ImmutableWarning />;
+  }
+
   return (
     <div
       className='container relative mx-auto my-8 max-w-6xl'
       id='repository-settings-section'
     >
-      {currentRepository?.is_immutable ? (
-        <p className='text-sm font-normal text-red-800 md:text-xl dark:text-red-400'>
-          {dict.repository.immutableDescription}
-        </p>
-      ) : (
-        <SettingsForm
-          initialValues={{
-            name: currentRepository?.name ?? '',
-            description: currentRepository?.description ?? '',
-            owner: currentRepository?.owner.id ?? '',
-          }}
-          onSubmit={handleUpdateRepository}
-          fieldConfiguration={fieldConfiguration}
-          deleteItem={handleDeleteRepository}
-          itemName='Repository'
-          submitButtonLabel={dict.repository.settings.saveChanges}
-          deleteButtonLabel={dict.repository.settings.deleteRepository}
-          dangerZoneMessage={dict.repository.settings.deletionNote}
-        />
-      )}
+      <SettingsForm
+        initialValues={{
+          name: currentRepository?.name ?? '',
+          description: currentRepository?.description ?? '',
+          owner: currentRepository?.owner.id ?? '',
+        }}
+        onSubmit={handleUpdateRepository}
+        fieldConfiguration={fieldConfiguration}
+        deleteItem={handleDeleteRepository}
+        itemName='Repository'
+        submitButtonLabel={dict.repository.settings.saveChanges}
+        deleteButtonLabel={dict.repository.settings.deleteRepository}
+        dangerZoneMessage={dict.repository.settings.deletionNote}
+      />
     </div>
   );
 };
